chaincfg: Move regnet code to regnet files.
This moves the various code that is specific to regnet to the
regnetparams.go file to keep the information housed together.  Along the
same lines, it moves the tests specific to the regnet params to
regnetparams_test.go.

It also unexports the block one ledger definition since it should only
be exposed via the params.
